+++ nix/overlays/haskell-overrides-next.nix
@@ -3,7 +3,7 @@ self: super:
 let
 
   inherit (self) haskell;
-  inherit (self.lib) withLocalPinPon withGitHubAmazonka;
+  inherit (self.lib) withLocalPinPon;
   inherit (haskell.lib) dontCheck doJailbreak noHaddocks;
 
   pinPonHlintPath = ../pkgs/pinpon-hlint.nix;
@@ -12,18 +12,5 @@ in
 {
 
   ## Testing with upcoming GHC releases.
-
-  ## GHC 8.4.2.
-
-  haskellPackages842 =
-    withGitHubAmazonka (withLocalPinPon pinPonHlintPath (self.haskell.packages.ghc842.extend (self: super:
-      rec {
-        http-media = doJailbreak super.http-media;
-        servant = doJailbreak super.servant;
-        servant-client = doJailbreak super.servant-client;
-        servant-server = doJailbreak super.servant-server;
-        servant-swagger-ui = doJailbreak super.servant-swagger-ui;
-      }
-    )));
 
 }